Ensemble training

+++ utils/gcforest/utils/win_utils.py
@@ -16,7 +16,7 @@ def get_windows_channel(X, X_win, des_id, nw, nh, win_x, win_y, stride_x, stride
     dj = des_id % win_x
     di = des_id / win_x % win_y
     k = des_id / win_x / win_y
-    src = X[:, k, di:di+nh*stride_y:stride_y, dj:dj+nw*stride_x:stride_x].ravel()
+    src = X[:, int(k), int(di):int(di+nh*stride_y):int(stride_y), int(dj):int(dj+nw*stride_x):int(stride_x)].ravel()
     des = X_win[des_id, :]
     np.copyto(des, src)
 
@@ -39,9 +39,9 @@ def get_windows(X, win_x, win_y, stride_x=1, stride_y=1, pad_x=0, pad_y=0):
         X = np.concatenate(( np.zeros((n, c, h, pad_x),dtype=X.dtype), X ), axis=3)
     n, c, h, w = X.shape
     nc = win_y * win_x * c
-    nh = (h - win_y) / stride_y + 1
-    nw = (w - win_x) / stride_x + 1
-    X_win = np.empty(( nc, n * nh * nw), dtype=np.float32)
+    nh = int((h - win_y) / stride_y + 1)
+    nw = int((w - win_x) / stride_x + 1)
+    X_win = np.empty(( nc, int(n * nh * nw)), dtype=np.float32)
     LOGGER.info("get_windows_start: X.shape={}, X_win.shape={}, nw={}, nh={}, c={}, win_x={}, win_y={}, stride_x={}, stride_y={}".format(
                 X.shape, X_win.shape, nw, nh, c, win_x, win_y, stride_x, stride_y))
     Parallel(n_jobs=-1, backend="threading", verbose=0)(
@@ -56,10 +56,10 @@ def calc_accuracy(y_gt, y_pred, tag):
     LOGGER.info("Accuracy({})={:.2f}%".format(tag, np.sum(y_gt==y_pred)*100./len(y_gt)))
 
 def win_vote(y_win_predict, n_classes):
-    """ 
-     
+    """
+
     y_win_predict (ndarray): n x n_window
-        y_win_predict[i, j] prediction for the ith data of jth window 
+        y_win_predict[i, j] prediction for the ith data of jth window
     """
     y_pred = np.zeros(len(y_win_predict), dtype=np.int16)
     for i, y_bag in enumerate(y_win_predict):
@@ -67,8 +67,8 @@ def win_vote(y_win_predict, n_classes):
     return y_pred
 
 def win_avg(y_win_proba):
-    """ 
-     
+    """
+
     Parameters
     ----------
     y_win_proba: n x n_windows x n_classes
